fix(app): выделен компонент App, использован тег main для главного контента
